requirements.txt
aiohttp==2.3.9
google-auth==1.3.0
# NOTE: temporary until we package & release gordon core
# gordon-dns==0.0.1.dev0
jsonschema==2.6.0
